+++ src/resources/views/channel/_list_for_card.blade.php
@@ -1,0 +1,19 @@
+|
+@if($model->channels->count())
+    {{ __('Channels') }}:
+    @can('view channels')
+        @foreach($model->channels->take(3) as $channel)
+            <a href="{{ route('vanilo.admin.channel.show', $channel->id) }}">
+                {{ $channel->name }}
+            </a>
+        @endforeach
+    @else
+        {{ $model->channels->take(3)->implode('name', ' | ') }}
+    @endcan
+@else
+    {{ __('no channels') }}
+@endif
+
+@if($model->channels->count() > 3)
+    | {{ __('+ :num more...', ['num' => $model->channels->count() - 3]) }}
+@endif