+++ src/resources/views/master-product/edit.blade.php
@@ -29,7 +29,9 @@
 
     <div class="col-12 col-lg-4 col-xl-3">
         @include('vanilo::media._edit', ['model' => $product])
-        @include('vanilo::channel._edit', ['model' => $product])
+        @if($multiChannelEnabled)
+            @include('vanilo::channel._edit', ['model' => $product])
+        @endif
     </div>
 </div>