+++ src/resources/views/payment-method/create.blade.php
@@ -5,8 +5,9 @@
 @stop
 
 @section('content')
-{!! Form::model($paymentMethod, ['route' => 'vanilo.admin.payment-method.store', 'autocomplete' => 'off']) !!}
+{!! Form::model($paymentMethod, ['route' => 'vanilo.admin.payment-method.store', 'autocomplete' => 'off', 'class' => 'row']) !!}
 
+<div class="col-12 col-lg-8 col-xl-9">
     <x-appshell::card accent="success">
         <x-slot:title>{{ __('Payment Method Details') }}</x-slot:title>
 
@@ -17,6 +18,13 @@
             <x-appshell::cancel-button />
         </x-slot:footer>
     </x-appshell::card>
+</div>
+
+<div class="col-12 col-lg-4 col-xl-3">
+    @if($multiChannelEnabled)
+        @include('vanilo::channel._create')
+    @endif
+</div>
 
 {!! Form::close() !!}
 @stop