+++ src/resources/views/payment-method/edit.blade.php
@@ -5,22 +5,35 @@
 @stop
 
 @section('content')
-{!! Form::model($paymentMethod, [
-        'route'  => ['vanilo.admin.payment-method.update', $paymentMethod],
-        'method' => 'PUT'
-    ])
-!!}
 
-    <x-appshell::card accent="secondary">
-        <x-slot:title>{{ __('Payment Method Details') }}</x-slot:title>
+<div class="row mb-4">
 
-        @include('vanilo::payment-method._form')
+    <div class="col-12 col-lg-8 col-xl-9 mb-4">
+        {!! Form::model($paymentMethod, [
+                'route'  => ['vanilo.admin.payment-method.update', $paymentMethod],
+                'method' => 'PUT'
+            ])
+        !!}
 
-        <x-slot:footer>
-            <x-appshell::save-button />
-            <x-appshell::cancel-button />
-        </x-slot:footer>
-    </x-appshell::card>
+        <x-appshell::card accent="secondary">
+            <x-slot:title>{{ __('Payment Method Details') }}</x-slot:title>
 
-{!! Form::close() !!}
+            @include('vanilo::payment-method._form')
+
+            <x-slot:footer>
+                <x-appshell::save-button />
+                <x-appshell::cancel-button />
+            </x-slot:footer>
+        </x-appshell::card>
+
+        {!! Form::close() !!}
+    </div>
+
+    <div class="col-12 col-lg-4 col-xl-3">
+        @if($multiChannelEnabled)
+            @include('vanilo::channel._edit', ['model' => $paymentMethod])
+        @endif
+    </div>
+</div>
+
 @stop