+++ src/resources/views/product/edit.blade.php
@@ -28,7 +28,9 @@
 
     <div class="col-12 col-lg-4 col-xl-3">
         @include('vanilo::media._edit', ['model' => $product])
-        @include('vanilo::channel._edit', ['model' => $product])
+        @if($multiChannelEnabled)
+            @include('vanilo::channel._edit', ['model' => $product])
+        @endif
     </div>
 
 </div>