+++ src/resources/views/product/show.blade.php
@@ -20,25 +20,7 @@
                 <x-slot:subtitle>
                     {{ $product->sku }}
                     @if($multiChannelEnabled)
-                        |
-                        @if($product->channels->count())
-                            {{ __('Channels') }}:
-                            @can('view channels')
-                                @foreach($product->channels->take(3) as $channel)
-                                    <a href="{{ route('vanilo.admin.channel.show', $channel->id) }}">
-                                        {{ $channel->name }}
-                                    </a>
-                                @endforeach
-                            @else
-                                {{ $product->channels->take(3)->implode('name', ' | ') }}
-                            @endcan
-                        @else
-                            {{ __('no channels') }}
-                        @endif
-
-                        @if($product->channels->count() > 3)
-                            | {{ __('+ :num more...', ['num' => $product->channels->count() - 3]) }}
-                        @endif
+                        @include('vanilo::channel._list_for_card', ['model' => $product])
                     @endif
                 </x-slot:subtitle>
             </x-appshell::card-with-icon>